Use dirname command in git hooks.
This follows how husky does it. I now know why it was done this way.

When the repository is moved, such as by copying, or by relocating it
to a remote server for execution, the absolute path that we currently
use in hooks will break.

It is not trivial to hardcode a relative path properly. According to
the documentation for hooks, the working directory prior to running
hooks is different depending on whether the repository is bare or not,
and it is different for certain client vs. server hooks.

We can sidestep these issues by using dirname.